changed build
It looks like the issue is still related to conflicting peer dependencies involving the @sanity/client package. This conflict is causing the npm install command to fail during the Vercel deployment.

One possible solution to resolve the issue is to use the --legacy-peer-deps flag when running npm install on Vercel. This flag allows npm to use an older, potentially less strict dependency resolution algorithm, which may help in resolving the conflicting peer dependencies.